+++ packages/two_dimensional_scrollables/CHANGELOG.md
@@ -1,3 +1,13 @@
+## NEXT
+
+## 0.3.2
+
+* Fixes a bug where the TreeView would not update correctly when the animation duration is zero.
+
+## 0.3.1
+
+* Adds generics to the callbacks and builders of TreeView.
+
 ## 0.3.0
 
 * Adds new TreeView widget and associated classes.